Do not recreate instances on changes of `image_id`
When referencing images by name outside the scope of the project, the
usual expectation is to not recreate resources when any of the attributes
change. E.g. when a cloud operator uploads new versions, thus creating
new IDs.
Changes to `image_id` should therefore be ignored and resource
replcement should be triggered only on explicit changes to the
referenced image name.

Signed-off-by: Jan Horstmann <horstmann@osism.tech>